Added save state on activity recreate and some fixes

+++ app/src/main/java/com/example/instattest/ApiServiceImpl.java
@@ -11,14 +11,14 @@ import retrofit2.Callback;
 import retrofit2.Retrofit;
 
 public class ApiServiceImpl {
-    public static void getMatch(int id, Callback<ResponseBody> callback) throws JSONException {
+    public static void getMatch(long id, Callback<ResponseBody> callback) throws JSONException {
         ApiService apiService = getApiService();
         JSONObject requestBody = new JSONObject("{\"proc\": \"get_match_info\", \"params\": {\"_p_sport\":1, \"_p_match_id\": " + id + "}}");
         Call<ResponseBody> getMatchCall = apiService.getMatch(RequestBody.create(MediaType.get("application/json"), requestBody.toString()));
         getMatchCall.enqueue(callback);
     }
 
-    public static void getVideos(int matchId, Callback<ResponseBody> callback) throws JSONException {
+    public static void getVideos(long matchId, Callback<ResponseBody> callback) throws JSONException {
         ApiService apiService = getApiService();
         JSONObject requestBody = new JSONObject("{\"match_id\": " + matchId + ",\"sport_id\":1}");
         Call<ResponseBody> getVideosCall = apiService.getVideos(RequestBody.create(MediaType.get("application/json"), requestBody.toString()));
